+++ JBQQuizMe.Repository/QuestionInfo.cs
@@ -1,8 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace JBQQuizMe.Repository
 {
@@ -10,6 +7,7 @@ namespace JBQQuizMe.Repository
     {
         public int Number { get; set; }
         public string Question { get; set; }
+        public int AnswerHash { get; set; }
         public List<string> Answer { get; set; }
         public List<List<string>>? WrongAnswers { get; set; }
         public string? Passage { get; set; }