+++ JBQQuizMeApp/Converters/TextLengthToFontSize.cs
@@ -9,27 +9,41 @@ namespace JBQQuizMeApp.Converters
             var strVal = (string)value;
             var strValLen = strVal.Length;
 
-            var fontSizeConverter = new FontSizeConverter();
-
+            decimal origSize;
             if (strValLen > 50)
             {
-                return fontSizeConverter.ConvertFromString("Micro");
-
+                origSize = 16m;
             }
             else if(strValLen > 25)
             {
-                return fontSizeConverter.ConvertFromString("Small");
-
+                origSize = 18m;
+            }
+            else if (strValLen > 10)
+            {
+                origSize = 22m;
             }
             else
             {
-                return fontSizeConverter.ConvertFromString("Medium");
+                origSize = 26m;
             }
+
+            return AdjustFontSize(origSize);
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
         {
             throw new NotImplementedException();
+        }
+
+        private decimal AdjustFontSize(decimal value)
+        {
+            if (DeviceInfo.Current.Platform == DevicePlatform.MacCatalyst)
+            {
+                // On MacOS on Catalyst the fonts are artificially 30% smaller, correct for this...
+                return value / 0.7m;
+            }
+
+            return value;
         }
     }
 }