+++ JBQQuizMeApp/QuizPage.xaml.cs
@@ -13,7 +13,7 @@ public partial class QuizPage : ContentPage
         context.ReadQuestions = readQuestions;
 
 #pragma warning disable CS4014 // Because this call is not awaited, execution of the current method continues before the call is completed
-        context.Initialize();
+        context.InitializeAsync();
 #pragma warning restore CS4014 // Because this call is not awaited, execution of the current method continues before the call is completed
     }
 }