+++ QuestionImporterApp/Program.cs
@@ -3,11 +3,15 @@ using QuestionImporterApp;
 using QuestionImporterApp.Validations;
 using System;
 using System.Globalization;
+using System.Security.Cryptography;
+using System.Text;
 
 namespace QuestionsImporterApp
 {
     internal class Program
     {
+        static private Dictionary<int, string> _stringHashes = new Dictionary<int, string>();
+
         static void Main(string[] args)
         {
             args = new string[] { "C:\\Users\\heali\\Documents\\GitHub\\JBQQuizMe\\JQB_10Pt_Questions.csv" };
@@ -56,35 +60,42 @@ namespace QuestionsImporterApp
                 {
                     Console.WriteLine("All validations passed");
 
+                    var key = GenerateKey();
+
+                    Console.WriteLine($"NEW KEY: {key}");
+
                     using StreamWriter file = new("question_data.cs.snippet", false);
 
                     foreach (var row in rows)
                     {
+                        var hasModifiedAnswer = !string.IsNullOrEmpty(row.Answer);
+
                         file.WriteLine(@"new QuestionInfo {");
                         file.WriteLine($"    Number = { row.Number },");
-                        file.WriteLine($"    Question = \"{ FormatString(row.Question) }\",");
-                        if(string.IsNullOrEmpty(row.Answer) == false)
+                        file.WriteLine($"    Question = \"{ EncryptString(key, row.Question) }\",");
+                        file.WriteLine($"    AnswerHash = { GetStringHash(hasModifiedAnswer ? row.AnswersAsList() : row.OriginalAnswersAsList()) },");
+                        if (hasModifiedAnswer)
                         {
                             if (row.IsAnswerList)
                             {
-                                var formattedList = row.AnswersAsList().Select(x => $"\"{ FormatString(x) }\"");
+                                var formattedList = row.AnswersAsList().Select(x => $"\"{ EncryptString(key, x) }\"");
                                 file.WriteLine($"    Answer = new List<string> {{ { string.Join(',', formattedList) } }},");
                             }
                             else
                             {
-                                file.WriteLine($"    Answer = new List<string> {{ \"{ FormatString(row.Answer) }\" }},");
+                                file.WriteLine($"    Answer = new List<string> {{ \"{ EncryptString(key, row.Answer) }\" }},");
                             }
                         }
                         else
                         {
                             if (row.IsOriginalAnswerList)
                             {
-                                var formattedList = row.OriginalAnswersAsList().Select(x => $"\"{ FormatString(x) }\"");
+                                var formattedList = row.OriginalAnswersAsList().Select(x => $"\"{ EncryptString(key, x) }\"");
                                 file.WriteLine($"    Answer = new List<string> {{ { string.Join(',', formattedList) } }},");
                             }
                             else
                             {
-                                file.WriteLine($"    Answer = new List<string> {{ \"{ FormatString(row.OriginalAnswer) }\" }},");
+                                file.WriteLine($"    Answer = new List<string> {{ \"{ EncryptString(key, row.OriginalAnswer) }\" }},");
                             }
                         }
 
@@ -98,7 +109,7 @@ namespace QuestionsImporterApp
                             file.WriteLine($"    WrongAnswers = new List<List<string>> {{");
                             foreach (var wrongAnswer in wrongAnswers)
                             {
-                                var formattedWrongAnswers = wrongAnswer.Split('~').Select(x => $"\"{ FormatString(x) }\"");
+                                var formattedWrongAnswers = wrongAnswer.Split('~').Select(x => $"\"{ EncryptString(key, x) }\"");
 
                                 file.WriteLine($"        new List<string> {{ {string.Join(',', formattedWrongAnswers)} }},");
                             }
@@ -133,9 +144,77 @@ namespace QuestionsImporterApp
             }
         }
 
-        static private string FormatString(string value)
+        static private string GenerateKey()
         {
-            return value.Replace(@"""", @"\""");
+            var crypto = Aes.Create();
+            crypto.KeySize = 128;
+            crypto.BlockSize = 128;
+            crypto.GenerateKey();
+            byte[] keyGenerated = crypto.Key;
+            return Convert.ToBase64String(keyGenerated);
+        }
+
+        static private string EncryptString(string key, string plainText)
+        {
+            byte[] iv = new byte[16];
+            byte[] array;
+
+            using (Aes aes = Aes.Create())
+            {
+                aes.Key = Encoding.UTF8.GetBytes(key);
+                aes.IV = iv;
+
+                ICryptoTransform encryptor = aes.CreateEncryptor(aes.Key, aes.IV);
+
+                using (var memoryStream = new MemoryStream())
+                {
+                    using (var cryptoStream = new CryptoStream(memoryStream, encryptor, CryptoStreamMode.Write))
+                    {
+                        using (var streamWriter = new StreamWriter(cryptoStream))
+                        {
+                            streamWriter.Write(plainText);
+                        }
+
+                        array = memoryStream.ToArray();
+                    }
+                }
+            }
+
+            return Convert.ToBase64String(array);
+        }
+
+        static private int GetStringHash(IEnumerable<string> values)
+        {
+            if (values == null)
+            {
+                throw new ArgumentNullException(nameof(values));
+            }
+            if (values.Count() == 0)
+            {
+                throw new ArgumentOutOfRangeException(nameof(values));
+            }
+            if (values.Any(x => string.IsNullOrEmpty(x)))
+            {
+                throw new ArgumentOutOfRangeException(nameof(values));
+            }
+
+            var mergedString = string.Join("|", values);
+
+            var hash = mergedString.GetHashCode();
+
+            if(_stringHashes.TryGetValue(hash, out var foundMergeString) == false)
+            {
+                _stringHashes.Add(hash, mergedString);
+            }
+            else
+            {
+                if (mergedString != foundMergeString)
+                {
+                    throw new InvalidOperationException("Deal with when/if it happens...");
+                }
+            }
+
+            return hash;
         }
     }
 }