+++ CHANGELOG.md
@@ -24,7 +24,7 @@ Track changes made to the game application.
 - Fix and clean up questions
 - Fix a number of build warnings
 
-## Version 0.4.0.0
+## Version 0.4.0.0 (app ver 1)
 
 - Second beta version
 - First Android build
@@ -36,7 +36,7 @@ Track changes made to the game application.
 - Further question clean up and add some more wrong answers
 - Fix start and end question range on main page
 
-## Version 0.5.0.0
+## Version 0.5.0.0 (app ver 2)
 
 - Third beta version
 - Lower lighting of candle from 20 to 10 correct answers
@@ -44,14 +44,14 @@ Track changes made to the game application.
 - Show annoying frog on wrong answers
 - Animate congratulation message when candle is fully lit
 
-## Version 0.6.0.0
+## Version 0.6.0.0 (app ver 3)
 
 - Fourth beta version
 - Add text to speech for questions and answers
 - Save text to speech settings between game runs
 - Save question start and end range between game runs
 
-## Version 0.7.0.0
+## Version 0.7.0.0 (app ver 4)
 
 - Fifth beta version
 - First MacOS build
@@ -59,3 +59,14 @@ Track changes made to the game application.
 - Add a cloud the player has to click away on a correct answer
 - Clean up SVG files
 - Known Issue: Font is smaller on MacOS (https://github.com/dotnet/maui/issues/10622)
+
+## Version 0.8.0.0 (app ver 5)
+
+- Sixth beta version
+- Pull code out of quiz view-model into 'question provider'
+- Fix issue where question is read while rain cloud is up
+- Implement proper UI styles
+- Address font size issue on MacOS in UI styles
+- Added an 'answer hash' to the question data to make generating wrong answers easier more efficient
+- Move decryption of data from provider into repository, which is more appropriate for a repository pattern
+- Add information about copyrighted material in the code